refactor(tui): split save preview pipeline
